better make to run cantas
mongodb and redis are necessary to run Cantas, especially in development. `make
run` will detect mongodb and redis and start them automatically if necessary on
behalf of you to give a convenient way. `make run` also give an example and
reference to developers to know what prerequisites are required to run Cantas.