make noreply email address and the email domain of invitation configurable
noreply email address is configurable in a simple way.

the domain name of invitation email is no longer limited to a specific domain.
Multiple domain names can be specified to `trusted_domains`. If there is no
limitation, '*' is the right one. Anyway, when '*' and common domain names are
mixed, no limitation there eventually.