load related modules only when specific auth method is enabled
Since Cantas supports several authentication strategies now and in the
future, most of the time, not all of them are enabled, in such a case
just load related dependent module for each enabled authentication
strategy is flexible and make it possible that administrator has the
opportunity to only install needed dependencies.

In particular, if someone deploys Cantas somewhere such as OpenShift
where it's impossible to `yum install` for Kerberos authentication, this
change can handle this situation easily.